Add windows permission usage script as a feature
